Move ZipInputPopUp from forecastChart to forecastCarousel
