+++ client/src/components/ForecastCarousel.js
@@ -1,13 +1,14 @@
 import { Carousel } from "react-responsive-carousel";
 import "react-responsive-carousel/lib/styles/carousel.min.css";
-import React from "react";
+import React, { useState, useEffect } from "react";
 import styled from "@emotion/styled";
 import ForecastChart from "./ForecastChart";
 import ForecastInfo from "./ForecastInfo";
+import { useParams } from "react-router-dom";
+import ZipInputPopUp from "./ZipInputPopUp";
 
 const CarouselContainer = styled(Carousel)`
   display: flex;
-  flex-grow: 1;
   li.slide {
     background: transparent;
   }
@@ -17,17 +18,37 @@ const CarouselContainer = styled(Carousel)`
   }
 `;
 
+const CarouselWrapper = styled.div`
+  /*   display: flex;
+  justify-content: center; */
+  flex-grow: 1;
+`;
+
 export const ForecastCarousel = () => {
+  const [isOpen, setIsOpen] = useState(false);
+  const { zip } = useParams();
+
+  useEffect(() => {
+    setIsOpen(false);
+  }, [zip]);
+
+  const togglePopup = () => {
+    setIsOpen(!isOpen);
+  };
+
   return (
-    <CarouselContainer
-      showArrows={false}
-      showStatus={false}
-      showThumbs={false}
-      swipeable={true}
-    >
-      <ForecastChart />
-      <ForecastInfo />
-    </CarouselContainer>
+    <CarouselWrapper>
+      {isOpen && <ZipInputPopUp handleClose={togglePopup} />}
+      <CarouselContainer
+        showArrows={false}
+        showStatus={false}
+        showThumbs={false}
+        swipeable={true}
+      >
+        <ForecastChart zipOnClick={togglePopup} />
+        <ForecastInfo />
+      </CarouselContainer>
+    </CarouselWrapper>
   );
 };
