+++ client/src/components/ForecastChart.js
@@ -1,18 +1,17 @@
 import React, { useState, useEffect } from "react";
 import ContentContainer from "./ContentContainer";
-import ZipInputPopUp from "./ZipInputPopUp";
 import { useParams } from "react-router-dom";
 import styled from "@emotion/styled";
 import Chart from "./Chart";
 import { fetchCity } from "../api/fetchForecast";
 import ZipButton from "./ZipButton";
+import PropTypes from "prop-types";
 
 const Headline = styled.h2`
   margin: 0.1rem;
 `;
 
-export const ForecastChart = () => {
-  const [isOpen, setIsOpen] = useState(false);
+export const ForecastChart = (props) => {
   const { zip } = useParams();
   const [cityName, setCityName] = useState(null);
 
@@ -25,22 +24,17 @@ export const ForecastChart = () => {
     doFetch();
   }, [zip]);
 
-  useEffect(() => {
-    setIsOpen(false);
-  }, [zip]);
-
-  const togglePopup = () => {
-    setIsOpen(!isOpen);
-  };
-
   return (
     <ContentContainer>
       <Headline>Ökostromanteil</Headline>
-      <ZipButton onClick={togglePopup}>{cityName}</ZipButton>
-      {isOpen && <ZipInputPopUp handleClose={togglePopup} />}
+      <ZipButton onClick={props.zipOnClick}>{cityName}</ZipButton>
       <Chart />
     </ContentContainer>
   );
 };
 
 export default ForecastChart;
+
+ForecastChart.propTypes = {
+  zipOnClick: PropTypes.func,
+};
